+++ NEWS.md
@@ -14,6 +14,7 @@ Version 28.1-dev (v28-fixes)
 
 * Autosave state whenever window becomes inactive, which retains
   the state when logging off on Gnome without closing the window
+* Fixed piece position after picking up piece from board in setup mode
 * Fixed workaround for QTBUG-99695 in Trigon, which sometimes causes
   wrong piece positions at the end of piece placement animations